requirements.txt
django
odoo-client-lib
python-decouple
django-widget-tweaks